Add files via upload
I think this will be the last version of the sensor part of the project. It is stable enough in my opinion. I will test it again when with the motors of the car and see if there is some imprecision. Noticed that it is one of my first electronic projects, so don't blame me for what you find in this code here 👀